- moved Sidepanel component to ui folder
- updated imports
