+++ frontend/src/components/ui/SidePanel/SidePanel.jsx
