+++ frontend/src/components/ui/SidePanel/SidePanel.module.scss
